boot: fix init_freemem skipping reserved regions
In cases where the kernel is passed a very fragmented map of
available memory (specifically when booting from GRUB on x86),
it is possible for multiple available regions to overlap with
a single reserved region.

Currently, if there is an overlap
(where avail_reg[a].start >= reserved[r].start), the available
region is trimmed and the region index is incremented. This causes
any other overlapping regions to be handled as if they are
non-overlapping, since it's now looking at the next reserved region.
This eventually leads to reserved memory being passed to the root
task and can cause corruption issues if it is used.

It is fine to skip this index increment, since the available region
is now either empty or non-overlapping, and both cases are already
trivially handled.

Signed-off-by: James Martin <fennelfoxxo@gmail.com>